+++ internal/service/commit/find_all_commits_test.go
@@ -49,6 +49,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 				"bf6e164cac2dc32b1f391ca4290badcbe4ffc5fb",
 				"48ca272b947f49eee601639d743784a176574a09",
 			},
+			BodySize: 48,
 		},
 		{
 			Id:        "48ca272b947f49eee601639d743784a176574a09",
@@ -57,6 +58,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320271),
 			Committer: dummyCommitAuthor(1500320271),
 			ParentIds: []string{"335bc94d5b7369b10251e612158da2e4a4aaa2a5"},
+			BodySize:  20,
 		},
 		{
 			Id:        "335bc94d5b7369b10251e612158da2e4a4aaa2a5",
@@ -65,6 +67,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320269),
 			Committer: dummyCommitAuthor(1500320269),
 			ParentIds: []string{"1039376155a0d507eba0ea95c29f8f5b983ea34b"},
+			BodySize:  20,
 		},
 	}
 
@@ -77,6 +80,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320272),
 			Committer: dummyCommitAuthor(1500320272),
 			ParentIds: []string{"9d526f87b82e2b2fd231ca44c95508e5e85624ca"},
+			BodySize:  11,
 		},
 		{
 			Id:        "9d526f87b82e2b2fd231ca44c95508e5e85624ca",
@@ -85,6 +89,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320270),
 			Committer: dummyCommitAuthor(1500320270),
 			ParentIds: []string{"1039376155a0d507eba0ea95c29f8f5b983ea34b"},
+			BodySize:  10,
 		},
 		{
 			Id:        "1039376155a0d507eba0ea95c29f8f5b983ea34b",
@@ -93,6 +98,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320268),
 			Committer: dummyCommitAuthor(1500320268),
 			ParentIds: []string{"54188278422b1fa877c2e71c4e37fc6640a58ad1"},
+			BodySize:  10,
 		}, {
 			Id:        "54188278422b1fa877c2e71c4e37fc6640a58ad1",
 			Subject:   []byte("Commit #7"),
@@ -100,6 +106,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320266),
 			Committer: dummyCommitAuthor(1500320266),
 			ParentIds: []string{"8b9270332688d58e25206601900ee5618fab2390"},
+			BodySize:  10,
 		}, {
 			Id:        "8b9270332688d58e25206601900ee5618fab2390",
 			Subject:   []byte("Commit #6"),
@@ -107,6 +114,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320264),
 			Committer: dummyCommitAuthor(1500320264),
 			ParentIds: []string{"f9220df47bce1530e90c189064d301bfc8ceb5ab"},
+			BodySize:  10,
 		}, {
 			Id:        "f9220df47bce1530e90c189064d301bfc8ceb5ab",
 			Subject:   []byte("Commit #5"),
@@ -114,6 +122,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320262),
 			Committer: dummyCommitAuthor(1500320262),
 			ParentIds: []string{"40d408f89c1fd26b7d02e891568f880afe06a9f8"},
+			BodySize:  10,
 		}, {
 			Id:        "40d408f89c1fd26b7d02e891568f880afe06a9f8",
 			Subject:   []byte("Commit #4"),
@@ -121,6 +130,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320260),
 			Committer: dummyCommitAuthor(1500320260),
 			ParentIds: []string{"df914c609a1e16d7d68e4a61777ff5d6f6b6fde3"},
+			BodySize:  10,
 		}, {
 			Id:        "df914c609a1e16d7d68e4a61777ff5d6f6b6fde3",
 			Subject:   []byte("Commit #3"),
@@ -128,6 +138,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320258),
 			Committer: dummyCommitAuthor(1500320258),
 			ParentIds: []string{"6762605237fc246ae146ac64ecb467f71d609120"},
+			BodySize:  10,
 		}, {
 			Id:        "6762605237fc246ae146ac64ecb467f71d609120",
 			Subject:   []byte("Commit #2"),
@@ -135,6 +146,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320256),
 			Committer: dummyCommitAuthor(1500320256),
 			ParentIds: []string{"79b06233d3dc769921576771a4e8bee4b439595d"},
+			BodySize:  10,
 		}, {
 			Id:        "79b06233d3dc769921576771a4e8bee4b439595d",
 			Subject:   []byte("Commit #1"),
@@ -142,6 +154,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500320254),
 			Committer: dummyCommitAuthor(1500320254),
 			ParentIds: []string{"1a0b36b3cdad1d2ee32457c102a8c0b7056fa863"},
+			BodySize:  10,
 		},
 		{
 			Id:      "1a0b36b3cdad1d2ee32457c102a8c0b7056fa863",
@@ -158,6 +171,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 				Date:  &timestamp.Timestamp{Seconds: 1393488198},
 			},
 			ParentIds: nil,
+			BodySize:  15,
 		},
 	}
 
@@ -170,6 +184,7 @@ func TestSuccessfulFindAllCommitsRequest(t *testing.T) {
 			Author:    dummyCommitAuthor(1500322381),
 			Committer: dummyCommitAuthor(1500322381),
 			ParentIds: []string{"1a0b36b3cdad1d2ee32457c102a8c0b7056fa863"},
+			BodySize:  11,
 		},
 	}