+++ internal/testhelper/commit.go
@@ -1,6 +1,7 @@
 package testhelper
 
 import (
+	"bytes"
 	"fmt"
 	"os"
 	"os/exec"
@@ -9,16 +10,45 @@ import (
 	"testing"
 )
 
+// CreateCommitOpts holds extra options for CreateCommit.
+type CreateCommitOpts struct {
+	Message  string
+	ParentID string
+}
+
 // CreateCommit makes a new empty commit and updates the named branch to point to it.
-func CreateCommit(t *testing.T, repoPath string, branchName string) string {
+func CreateCommit(t *testing.T, repoPath, branchName string, opts *CreateCommitOpts) string {
+
+	message := "message"
 	// The ID of an arbitrary commit known to exist in the test repository.
 	parentID := "1a0b36b3cdad1d2ee32457c102a8c0b7056fa863"
+
+	if opts != nil {
+		if opts.Message != "" {
+			message = opts.Message
+		}
+
+		if opts.ParentID != "" {
+			parentID = opts.ParentID
+		}
+	}
+
+	committerName := "Scrooge McDuck"
+	committerEmail := "scrooge@mcduck.com"
+
+	// message can be very large, passing it directly in args would blow things up!
+	stdin := bytes.NewBufferString(message)
 
 	// Use 'commit-tree' instead of 'commit' because we are in a bare
 	// repository. What we do here is the same as "commit -m message
 	// --allow-empty".
-	commitArgs := []string{"-C", repoPath, "commit-tree", "-m", "message", "-p", parentID, parentID + "^{tree}"}
-	newCommit := MustRunCommand(t, nil, "git", commitArgs...)
+	commitArgs := []string{
+		"-c", fmt.Sprintf("user.name=%s", committerName),
+		"-c", fmt.Sprintf("user.email=%s", committerEmail),
+		"-C", repoPath,
+		"commit-tree", "-F", "-", "-p", parentID, parentID + "^{tree}",
+	}
+	newCommit := MustRunCommand(t, stdin, "git", commitArgs...)
 	newCommitID := strings.TrimSpace(string(newCommit))
 
 	MustRunCommand(t, nil, "git", "-C", repoPath, "update-ref", "refs/heads/"+branchName, newCommitID)